Update a type alias in shared.h
Updated the type for `score_t` to `uint32_t` allocating 4 bytes of
memory for score storage as 2 bytes was quite low for higher game
scores.

+++ src/include/shared.h
@@ -10,7 +10,7 @@
 typedef uint16_t cell_t;
 typedef uint16_t input_t;
 typedef uint8_t select_t;
-typedef uint16_t score_t;
+typedef uint32_t score_t;
 typedef int8_t index_t;
 typedef uint16_t pos_t;
 typedef uint16_t len_t;
